requirements.txt
matplotlib
numpy
rasterio
svg.path
scipy
shapely
tqdm
networkx